+++ Frontend-LookStyle/app/home/Adaptability.tsx
