+++ Frontend-LookStyle/app/home/Guide.tsx
