+++ Frontend-LookStyle/app/home/Hero.tsx
