+++ Frontend-LookStyle/app/home/NavBar.tsx
